Added popup in all pages
